Removed the "location address" parameter from the non-UI layers
This address just derives from the location, so it was essentially a duplicate. The interface really didn't need this concept. Also, we were passing around "Obtaining location" instead of the actual address in the case of failures, which is wrong. Last but not least, for the user we'd rather have some other indication of progress.