Adjusted dependency transitivity on a per-cases basis
For some, transitivity is required for the clients to be able to use the objects in the responses, so I restored it. For some others, we keep it explicitely blocked (like for the wrapper implementations). And, for others, like the commons and the model, we leave it true, since we agree they can be used from anywhere.